requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.9.1
requests==2.31.0 